+++ packages/analytics-core/src/utils/debug.js
@@ -1,12 +1,10 @@
-import { set, globalContext } from '@analytics/global-storage-utils'
+import { set, globalContext, KEY } from '@analytics/global-storage-utils'
 import { compose } from '../vendor/redux'
-import { LIBRARY_NAME, PREFIX } from './internalConstants'
+import { LIB_NAME } from './internalConstants'
 
 export function Debug() {
-  // initialize global history
-  const globalVariable = PREFIX + LIBRARY_NAME + PREFIX
-  // Global key is window.__analytics__
-  set(globalVariable, [])
+  // Global key is window.__global__.analytics
+  set(LIB_NAME, [])
   // Return debugger
   return (createStore) => {
     return (reducer, preloadedState, enhancer) => {
@@ -14,7 +12,7 @@ export function Debug() {
       const origDispatch = store.dispatch
       const dispatch = (action) => {
         const a = action.action || action
-        globalContext[globalVariable].push(a)
+        globalContext[KEY][LIB_NAME].push(a)
         return origDispatch(action)
       }
       return Object.assign(store, { dispatch: dispatch })