+++ packages/analytics-util-router/src/index.js
@@ -1,37 +1,34 @@
 
 /* Listener for pushstate route changes */
-export default function routeChange(callback) {
-  if (typeof document === 'undefined') return
+export default function onRouteChange(fn) {
+  if (typeof window === 'undefined') return
   const { addEventListener, history, location } = window
   // Set initial pathname
-  let previousPath = location.pathname
+  let previous = location.pathname
   // Observe native navigation
-  addEventListener('popstate', () => {
-    callback(location.pathname)
-  })
+  addEventListener('popstate', () => fn(location.pathname))
   // Observe manual navigation
-  const methods = ['push', 'replace']
-  methods.map((type) => {
+  ;['push', 'replace'].map((type) => {
     const state = `${type}State`
-    const historyState = history[state]
-    history[state] = function () {
-      const updatedPath = trimPath(arguments[2])
-      if (previousPath !== updatedPath) {
-        // Set updated pathname
-        previousPath = updatedPath
-        // Debounce delay for SPA
-        setTimeout(() => callback(arguments[2]), 0)
+    const historyFn = history[state]
+    history[state] = function() {
+      const updated = trimPath(arguments[2])
+      if (previous !== updated) {
+        previous = updated // Set updated pathname
+        setTimeout(() => fn(arguments[2]), 0) // Debounce delay for SPA
       }
-      return historyState.apply(history, arguments)
+      return historyFn.apply(history, arguments)
     }
   })
 }
 
+function remove(str, char) {
+  const has = str.indexOf(char)
+  return (has > -1) ? str.slice(0, has) : str
+}
+
 // Trim hash and params off parcial url path
 function trimPath(urlPath) {
-  const hasParam = urlPath.indexOf('?')
-  if (hasParam > -1) urlPath = urlPath.slice(0, hasParam)
-  const hasHash = urlPath.indexOf('#')
-  if (hasHash > -1) urlPath = urlPath.slice(0, hasHash)
+  ['#', '?'].forEach((x) => urlPath = remove(urlPath, x))
   return urlPath
 }