+++ packages/analytics-plugin-amplitude/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/DavidWells/analytics/compare/@analytics/amplitude@0.1.2...@analytics/amplitude@0.1.3) (2021-10-29)
+
+**Note:** Version bump only for package @analytics/amplitude
+
+
+
+
+
+## [0.1.2](https://github.com/DavidWells/analytics/compare/@analytics/amplitude@0.1.1...@analytics/amplitude@0.1.2) (2021-10-24)
+
+
+### Bug Fixes
+
+* integrity sha ([70e7c1d](https://github.com/DavidWells/analytics/commit/70e7c1d))
+
+
+
+
+
+## [0.1.1](https://github.com/DavidWells/analytics/compare/@analytics/amplitude@0.1.0...@analytics/amplitude@0.1.1) (2021-10-24)
+
+
+### Bug Fixes
+
+* iife semicolon ([030af04](https://github.com/DavidWells/analytics/commit/030af04))
+
+
+
+
+
 # [0.1.0](https://github.com/DavidWells/analytics/compare/@analytics/amplitude@0.0.4...@analytics/amplitude@0.1.0) (2021-10-24)