+++ packages/analytics-plugin-aws-pinpoint/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.5](https://github.com/DavidWells/analytics/compare/@analytics/aws-pinpoint@0.7.4...@analytics/aws-pinpoint@0.7.5) (2022-02-05)
+
+**Note:** Version bump only for package @analytics/aws-pinpoint
+
+
+
+
+
+## [0.7.4](https://github.com/DavidWells/analytics/compare/@analytics/aws-pinpoint@0.7.3...@analytics/aws-pinpoint@0.7.4) (2022-01-03)
+
+**Note:** Version bump only for package @analytics/aws-pinpoint
+
+
+
+
+
+## [0.7.3](https://github.com/DavidWells/analytics/compare/@analytics/aws-pinpoint@0.7.2...@analytics/aws-pinpoint@0.7.3) (2022-01-02)
+
+**Note:** Version bump only for package @analytics/aws-pinpoint
+
+
+
+
+
+## [0.7.2](https://github.com/DavidWells/analytics/compare/@analytics/aws-pinpoint@0.7.1...@analytics/aws-pinpoint@0.7.2) (2021-12-12)
+
+**Note:** Version bump only for package @analytics/aws-pinpoint
+
+
+
+
+
 ## [0.7.1](https://github.com/DavidWells/analytics/compare/@analytics/aws-pinpoint@0.7.0...@analytics/aws-pinpoint@0.7.1) (2021-10-24)
 
 **Note:** Version bump only for package @analytics/aws-pinpoint