+++ packages/analytics-util-activity/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.12](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-activity/compare/@analytics/activity-utils@0.1.11...@analytics/activity-utils@0.1.12) (2022-02-05)
+
+**Note:** Version bump only for package @analytics/activity-utils
+
+
+
+
+
+## [0.1.11](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-activity/compare/@analytics/activity-utils@0.1.10...@analytics/activity-utils@0.1.11) (2022-01-03)
+
+**Note:** Version bump only for package @analytics/activity-utils
+
+
+
+
+
+## [0.1.10](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-activity/compare/@analytics/activity-utils@0.1.9...@analytics/activity-utils@0.1.10) (2022-01-02)
+
+**Note:** Version bump only for package @analytics/activity-utils
+
+
+
+
+
+## [0.1.9](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-activity/compare/@analytics/activity-utils@0.1.8...@analytics/activity-utils@0.1.9) (2021-12-12)
+
+**Note:** Version bump only for package @analytics/activity-utils
+
+
+
+
+
 ## [0.1.8](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-activity/compare/@analytics/activity-utils@0.1.7...@analytics/activity-utils@0.1.8) (2021-10-24)
 
 **Note:** Version bump only for package @analytics/activity-utils