+++ packages/analytics-util-redact/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.7](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-redact/compare/@analytics/redact-utils@0.0.6...@analytics/redact-utils@0.0.7) (2022-02-05)
+
+**Note:** Version bump only for package @analytics/redact-utils
+
+
+
+
+
+## [0.0.6](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-redact/compare/@analytics/redact-utils@0.0.5...@analytics/redact-utils@0.0.6) (2022-01-03)
+
+**Note:** Version bump only for package @analytics/redact-utils
+
+
+
+
+
+## [0.0.5](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-redact/compare/@analytics/redact-utils@0.0.4...@analytics/redact-utils@0.0.5) (2022-01-02)
+
+**Note:** Version bump only for package @analytics/redact-utils
+
+
+
+
+
+## [0.0.4](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-redact/compare/@analytics/redact-utils@0.0.3...@analytics/redact-utils@0.0.4) (2021-12-12)
+
+**Note:** Version bump only for package @analytics/redact-utils
+
+
+
+
+
 ## [0.0.3](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-redact/compare/@analytics/redact-utils@0.0.2...@analytics/redact-utils@0.0.3) (2021-10-24)
 
 **Note:** Version bump only for package @analytics/redact-utils