+++ packages/analytics-util-storage-global/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-global-storage/compare/@analytics/global-storage-utils@0.1.3...@analytics/global-storage-utils@0.1.4) (2022-02-05)
+
+**Note:** Version bump only for package @analytics/global-storage-utils
+
+
+
+
+
+## [0.1.3](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-global-storage/compare/@analytics/global-storage-utils@0.1.2...@analytics/global-storage-utils@0.1.3) (2021-12-12)
+
+**Note:** Version bump only for package @analytics/global-storage-utils
+
+
+
+
+
 ## [0.1.2](https://github.com/DavidWells/analytics/tree/master/packages/analytics-util-global-storage/compare/@analytics/global-storage-utils@0.1.1...@analytics/global-storage-utils@0.1.2) (2021-07-28)
 
 **Note:** Version bump only for package @analytics/global-storage-utils